Refactor entrapment_analyses into organized EntrapmentAnalysis module
Major changes:
- Created proper Julia module structure with src/, test/, and examples/ directories
- Split functionality into logical components:
  - core/efdr_methods.jl: EFDR calculation methods with generic integer support
  - core/entrapment_pairing.jl: Pairing algorithms
  - core/scoring.jl: Score manipulation functions
  - analysis/: EFDR analysis and calibration functions
  - plotting/: Visualization functions
  - api.jl: Main entry point with run_efdr_analysis()

Bug fixes:
- Fixed DataFrame indexing from deprecated df[:col] to df[\!, :col]
- Fixed map operation to work on column instead of entire DataFrame
- Made EFDR structs generic over integer types to accept UInt8 labels
- Added missing add_efdr_columns\! function

New features:
- Command-line interface via run_analysis.jl
- Comprehensive README documentation
- Example scripts for basic and real-world usage
- Organized test suite with proper test runner

Breaking changes:
- Removed old unstructured files (entrapment.jl, efdr_funcs.jl, etc.)
- Module now requires: using EntrapmentAnalysis
