--- test/entrapment_analyses/efdr_out/calibration_global_combined.csv
@@ -1,21 +1,0 @@
-bin,estimated_efdr,actual_fdr,error,n_items
-1,0.0,0.0,0.0,6470
-2,0.0,0.0,0.0,6470
-3,0.0,0.0,0.0,6470
-4,0.0,0.0,0.0,6470
-5,0.0,0.0,0.0,6470
-6,0.0,0.0,0.0,6470
-7,0.0,0.0,0.0,6470
-8,0.0,0.0,0.0,6470
-9,5.195765849576307e-5,3.434655675768504e-5,1.761110173807803e-5,6470
-10,0.000164718686300009,7.727975270479135e-5,8.743893359521764e-5,6470
-11,0.00021366091969917918,0.00011240691302515104,0.00010125400667402814,6470
-12,0.00034927629720407226,0.00025759917568263783,9.167712152143443e-5,6470
-13,0.000618015526357808,0.00036856497443823566,0.00024945055191957235,6470
-14,0.0008982791372693466,0.0005740781629498785,0.00032420097431946806,6470
-15,0.0012915322470777684,0.0007521895929933025,0.0005393426540844659,6470
-16,0.001987928859929203,0.0011495363214837713,0.0008383925384454316,6470
-17,0.0027216513741632788,0.0017001545595054096,0.0010214968146578692,6470
-18,0.004078936782248386,0.0025588184784475355,0.0015201183038008504,6470
-19,0.006480591539323927,0.003847718213617506,0.0026328733257064206,6470
-20,0.009707060861563048,0.005989180834621329,0.0037178800269417195,6470
