--- test/entrapment_analyses/efdr_out/calibration_global_paired.csv
@@ -1,21 +1,0 @@
-bin,estimated_efdr,actual_fdr,error,n_items
-1,0.0,0.0,0.0,6470
-2,0.0,0.0,0.0,6470
-3,0.0,0.0,0.0,6470
-4,0.0,0.0,0.0,6470
-5,0.0,0.0,0.0,6470
-6,0.0,0.0,0.0,6470
-7,0.0,0.0,0.0,6470
-8,0.0,0.0,0.0,6470
-9,5.1957510442347194e-5,3.434655675768504e-5,1.7610953684662155e-5,6470
-10,0.00020006865809583238,7.727975270479135e-5,0.00012278890539104103,6470
-11,0.00027394974743562284,0.00011240691302515104,0.0001615428344104718,6470
-12,0.0004458094731934228,0.00025759917568263783,0.00018821029751078496,6470
-13,0.0007658506797747332,0.00036856497443823566,0.0003972857053364976,6470
-14,0.001038158622248116,0.0005740781629498785,0.00046408045929823753,6470
-15,0.0014649530390610664,0.0007521895929933025,0.0007127634460677639,6470
-16,0.002233284742777019,0.0011495363214837713,0.0010837484212932477,6470
-17,0.0031118182026523273,0.0017001545595054096,0.0014116636431469177,6470
-18,0.0047073017493103365,0.0025588184784475355,0.002148483270862801,6470
-19,0.007452475652201969,0.003847718213617506,0.0036047574385844627,6470
-20,0.011175174778203375,0.005989180834621329,0.005185993943582046,6470
