--- test/entrapment_analyses/efdr_out/calibration_local_combined.csv
@@ -1,21 +1,0 @@
-bin,estimated_efdr,actual_fdr,error,n_items
-1,0.0,0.0,0.0,6470
-2,0.0,0.0,0.0,6470
-3,0.0,0.0,0.0,6470
-4,8.788189403426151e-5,0.00011591962905718702,2.8037735022925516e-5,6470
-5,0.0005996033384040083,0.00043276661514683154,0.0001668367232571767,6470
-6,0.0017218114352658164,0.0009015971148892323,0.0008202143203765841,6470
-7,0.001942361636837963,0.0008390373150805918,0.0011033243217573713,6470
-8,0.0019639863096240196,0.0008693972179289026,0.001094589091695117,6470
-9,0.002020013674520133,0.0009788768675940238,0.0010411368069261093,6470
-10,0.0020677279632103094,0.0009737248840803709,0.0010940030791299384,6470
-11,0.0021055008622106993,0.0010116622172263595,0.0010938386449843398,6470
-12,0.0021631402079542363,0.0010947964966512107,0.0010683437113030256,6470
-13,0.0023412305373022866,0.0012126976578290334,0.0011285328794732531,6470
-14,0.0026236327146907613,0.0013910355486862441,0.0012325971660045172,6470
-15,0.0030170345216937447,0.001586810922205049,0.0014302235994886957,6470
-16,0.0035261247128162984,0.0019802936630602783,0.0015458310497560202,6470
-17,0.004534012646270656,0.002609328120738249,0.0019246845255324074,6470
-18,0.0059581217623757765,0.003391722479821398,0.0025663992825543784,6470
-19,0.007949710252888879,0.004539168632555113,0.0034105416203337657,6470
-20,0.01049078708911287,0.005989180834621329,0.004501606254491541,6470
