--- test/entrapment_analyses/efdr_out/calibration_local_paired.csv
@@ -1,21 +1,0 @@
-bin,estimated_efdr,actual_fdr,error,n_items
-1,0.0,0.0,0.0,6470
-2,0.0,0.0,0.0,6470
-3,0.0,0.0,0.0,6470
-4,8.788164361468944e-5,0.00011591962905718702,2.8037985442497585e-5,6470
-5,0.0006163457828641422,0.00043276661514683154,0.0001835791677173107,6470
-6,0.0018254393493644285,0.0009015971148892323,0.0009238422344751962,6470
-7,0.002079921195559068,0.0008831971737690439,0.001196724021790024,6470
-8,0.0021088880454811104,0.0008887171561051004,0.00122017088937601,6470
-9,0.0021712759692774923,0.0009788768675940238,0.0011923991016834685,6470
-10,0.002233867670968923,0.0009582689335394126,0.0012755987374295102,6470
-11,0.0022688271317793415,0.0010116622172263595,0.001257164914552982,6470
-12,0.002365112237092729,0.0010947964966512107,0.0012703157404415184,6470
-13,0.0025990165240343745,0.0012008084651052194,0.0013982080589291551,6470
-14,0.0028909165735598557,0.0013910355486862441,0.0014998810248736116,6470
-15,0.00333622925948893,0.0015971148892323544,0.0017391143702565756,6470
-16,0.003951507550384629,0.0019802936630602783,0.001971213887324351,6470
-17,0.00510215834726316,0.002609328120738249,0.002492830226524911,6470
-18,0.006707450834994903,0.003391722479821398,0.003315728355173505,6470
-19,0.009051717686259265,0.004539168632555113,0.0045125490537041525,6470
-20,0.011998508042823191,0.005989180834621329,0.006009327208201862,6470
