--- test/entrapment_analyses/efdr_out/efdr_comparison_global.csv
@@ -1,4 +1,0 @@
-threshold,qval_n,qval_actual_fdr,combined_n,combined_efdr,paired_n,paired_efdr
-0.001,98162,0.0008659155273934924,98162,0.0017318359896559038,98162,0.001966137609258165
-0.01,129406,0.0059966307590065375,129406,0.011993944485186384,129406,0.013794542461688266
-0.05,129406,0.0059966307590065375,129406,0.011993944485186384,129406,0.013794542461688266
