--- test/entrapment_analyses/efdr_out/efdr_comparison_local.csv
@@ -1,4 +1,0 @@
-threshold,qval_n,qval_actual_fdr,combined_n,combined_efdr,paired_n,paired_efdr
-0.001,36463,0.0010421523187889093,36463,0.002084310576823535,36463,0.0022214299426816225
-0.01,129406,0.0059966307590065375,129406,0.011993295692924849,129406,0.013739702950404154
-0.05,129406,0.0059966307590065375,129406,0.011993295692924849,129406,0.013739702950404154
